Moving all Ruby Projects to a seperate git repo
All the projects that was done for The Odin Project is under this repository.
This repo will be used to maintain and display the progress for all the projects that are in the course.
For further information refer to the original theOdinProject repository. (If you're not me, drop me a message)